backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
django-filter==23.3
django-storages==1.14.2
django-markdownx==4.0.6
django-taggit==4.0.0
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
celery==5.3.5
redis==5.0.1
dj-database-url==2.1.0
whitenoise==6.6.0
django-environ==0.11.2
django-admin-interface==0.30.0
django-colorfield==0.14.0
django-summernote 